♻️ Change searchpage in a component JobSearchSection
